replace firebase and i18 files to config folders
