+++ src/assets/images/logoQuiz.png
