+++ src/config/i18/i18nextInit.js
@@ -2,9 +2,9 @@ import i18n from "i18next";
 import { initReactI18next } from "react-i18next";
 import Backend from "i18next-http-backend";
 import LanguageDetector from "i18next-browser-languagedetector";
-import translationRU from "./assets/locales/ru/translation.json";
-import translationEN from "./assets/locales/en/translation.json";
-import translationDEU from "./assets/locales/deu/translation.json";
+import translationRU from "../../assets/locales/ru/translation.json";
+import translationEN from "../../assets/locales/en/translation.json";
+import translationDEU from "../../assets/locales/deu/translation.json";
 
 const fallbackLng = ["ru"];
 const availableLanguages = ["ru", "en", "deu"];
