+++ src/modules/UserForm/api/sendUserInfoDb.js
@@ -1,6 +1,6 @@
 import { update } from "firebase/database";
 import { ref } from "firebase/database";
-import {db} from "../../../index";
+import db from "../../../config/firebase/firebaseConfig";
 
 const sendUserInfoDB = async (idUser) => {
     try {
